+++ yarpctest/fake_transport.go
@@ -21,6 +21,8 @@
 package yarpctest
 
 import (
+	"fmt"
+
 	"go.uber.org/yarpc/api/peer"
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/peer/hostport"
@@ -42,6 +44,7 @@ func NopTransportOption(nopOption string) FakeTransportOption {
 func NewFakeTransport(opts ...FakeTransportOption) *FakeTransport {
 	t := &FakeTransport{
 		Lifecycle: lifecycletest.NewNop(),
+		peers:     make(map[string]*FakePeer),
 	}
 	for _, opt := range opts {
 		opt(t)
@@ -53,6 +56,7 @@ func NewFakeTransport(opts ...FakeTransportOption) *FakeTransport {
 type FakeTransport struct {
 	transport.Lifecycle
 	nopOption string
+	peers     map[string]*FakePeer
 }
 
 // NopOption returns the configured nopOption. It's fake.
@@ -60,12 +64,59 @@ func (t *FakeTransport) NopOption() string {
 	return t.nopOption
 }
 
+// SimulateConnect simulates a connection to the peer, marking the peer as
+// available and notifying subscribers.
+func (t *FakeTransport) SimulateConnect(id peer.Identifier) {
+	t.Peer(id).simulateConnect()
+}
+
+// SimulateConnect simulates a disconnection to the peer, marking the peer as
+// unavailable and notifying subscribers.
+func (t *FakeTransport) SimulateDisconnect(id peer.Identifier) {
+	t.Peer(id).simulateDisconnect()
+}
+
+// Peer returns the persistent peer object for that peer identifier for the
+// lifetime of the fake transport.
+func (t *FakeTransport) Peer(id peer.Identifier) *FakePeer {
+	if peer, ok := t.peers[id.Identifier()]; ok {
+		return peer
+	} else {
+		peer := &FakePeer{id: id.(hostport.PeerIdentifier)}
+		t.peers[id.Identifier()] = peer
+		return peer
+	}
+}
+
 // RetainPeer returns a fake peer.
 func (t *FakeTransport) RetainPeer(id peer.Identifier, ps peer.Subscriber) (peer.Peer, error) {
-	return &FakePeer{id: id.(hostport.PeerIdentifier)}, nil
+	peer := t.Peer(id)
+	peer.subscribers = append(peer.subscribers, ps)
+	return peer, nil
 }
 
 // ReleasePeer does nothing.
 func (t *FakeTransport) ReleasePeer(id peer.Identifier, ps peer.Subscriber) error {
+	peer := t.Peer(id)
+	if subscribers, count := filterSubscriber(peer.subscribers, ps); count == 0 {
+		return fmt.Errorf("no such subscriber")
+	} else if count > 1 {
+		return fmt.Errorf("extra subscribers: %d", count-1)
+	} else {
+		peer.subscribers = subscribers
+	}
 	return nil
+}
+
+func filterSubscriber(subs []peer.Subscriber, ps peer.Subscriber) ([]peer.Subscriber, int) {
+	res := make([]peer.Subscriber, 0, len(subs))
+	count := 0
+	for _, sub := range subs {
+		if sub != ps {
+			res = append(res, sub)
+		} else {
+			count++
+		}
+	}
+	return res, count
 }